INS-3287: extract slow log test on env variable
